test: 딜러 테스트 불필요한 코드 제거
