+++ src/test/java/blackjack/gamer/DealerTest.java
@@ -1,13 +1,16 @@
 package blackjack.gamer;
 
 import static blackjack.fixture.TestFixture.provideBlackjackCards;
+import static blackjack.fixture.TestFixture.provideSum15Cards;
 import static org.assertj.core.api.Assertions.assertThat;
 
 import blackjack.card.Card;
+import blackjack.card.Denomination;
 import blackjack.card.Shape;
+import blackjack.cardMachine.CardMachine;
 import blackjack.cardMachine.CardRandomMachine;
-import blackjack.fixture.TestFixture.TestCardRandomMachine;
 import java.util.ArrayList;
+import java.util.Collections;
 import java.util.HashSet;
 import java.util.List;
 import java.util.Set;
@@ -15,6 +18,38 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
 class DealerTest {
+
+    public static class TestCardRandomMachine implements CardMachine {
+
+        private final List<Card> deck = List.of(
+                new Card(Shape.CLOB, Denomination.ACE),
+                new Card(Shape.CLOB, Denomination.TWO),
+                new Card(Shape.CLOB, Denomination.THREE)
+        );
+
+        @Override
+        public void receiveDeck(final List<Card> deck) {
+
+        }
+
+        @Override
+        public Card drawOneCard() {
+            final List<Card> mutableDeck = new ArrayList<>(deck);
+            Collections.shuffle(mutableDeck);
+            return mutableDeck.getFirst();
+        }
+
+    }
+
+    @DisplayName("카드 한 장을 랜덤하게 뽑는다.")
+    @Test
+    void SpreadOneRandomCard() {
+        // given
+        final Dealer dealer = new Dealer(new TestCardRandomMachine());
+
+        // when & then
+        assertThat(dealer.spreadOneCard()).extracting("shape").isEqualTo(Shape.CLOB);
+    }
 
     @DisplayName("52장이 곂치지 않게 뽑힌다.")
     @Test
@@ -34,16 +69,6 @@ class DealerTest {
         assertThat(uniqueCards.size() == deck.size()).isTrue();
     }
 
-    @DisplayName("카드 한 장을 랜덤하게 뽑는다.")
-    @Test
-    void SpreadOneRandomCard() {
-        // given
-        final Dealer dealer = new Dealer(new TestCardRandomMachine());
-
-        // when & then
-        assertThat(dealer.spreadOneCard()).extracting("shape").isEqualTo(Shape.CLOB);
-    }
-
     @DisplayName("딜러가 2장의 카드를 뽑는다.")
     @Test
     void spreadTwoCards() {
@@ -61,11 +86,10 @@ class DealerTest {
     @Test
     void isHit() {
         // given
-        final Dealer dealer = new Dealer(new TestCardRandomMachine());
+        final Dealer dealer = new Dealer(new CardRandomMachine());
 
         // when
-        final List<Card> cards = dealer.spreadTwoCards();
-        dealer.receiveCards(cards);
+        dealer.receiveCards(provideSum15Cards());
 
         // then
         assertThat(dealer.isHit()).isTrue();
@@ -73,7 +97,7 @@ class DealerTest {
 
     @DisplayName("딜러가 번 돈을 확인한다.")
     @Test
-    void updateEarnenMoney() {
+    void updateEarnedMoney() {
         // given
         final Dealer dealer = new Dealer(new CardRandomMachine());
 
@@ -113,11 +137,10 @@ class DealerTest {
     @Test
     void isBust() {
         // given
-        final Dealer dealer = new Dealer(new TestCardRandomMachine());
+        final Dealer dealer = new Dealer(new CardRandomMachine());
 
         // when
-        final List<Card> cards = dealer.spreadTwoCards();
-        dealer.receiveCards(cards);
+        dealer.receiveCards(provideSum15Cards());
 
         // then
         assertThat(dealer.isBust(16)).isFalse();
@@ -127,11 +150,10 @@ class DealerTest {
     @Test
     void isBlackjack() {
         // given
-        final Dealer dealer = new Dealer(new TestCardRandomMachine());
+        final Dealer dealer = new Dealer(new CardRandomMachine());
 
         // when
-        final List<Card> cards = provideBlackjackCards();
-        dealer.receiveCards(cards);
+        dealer.receiveCards(provideBlackjackCards());
 
         // then
         assertThat(dealer.isBlackjack(21, 2)).isTrue();
